move sip-031 types and functions to stacks-common
